Disconnect the Strip objects from the mapping
Strip objects often get discarded and regenerated due to our lazy
allocation strategy, so holding them in the mapping is a bad plan-
especially now that we have copy-on-mutation structures in the
registry.

Instead of doing that, we store the MAC address of the holding pusher
and the strip number, and look them up on each access-  this is not as
slow as it sounds as it's just a map lookup.